# webcalendar-openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WebCalendar API
  description: REST API for WebCalendar Next Generation (WCTNG)
  version: 2.0.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:47180/api/v2
    description: Local development
  - url: https://api.example.com/api/v2
    description: Production (placeholder)
...
security:
  - bearerAuth: []
paths:
...
  /events:
    get:
      operationId: listEvents
      summary: List events in a date range
      description: Returns events for the authenticated user within the specified date range, with pagination.
      tags:
        - Events
      parameters:
        - name: start
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{8}$'
          description: Start date (YYYYMMDD)
          example: "20260301"
        - name: end
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{8}$'
          description: End date (YYYYMMDD)
          example: "20260331"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Items per page
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
                  error:
                    type: "null"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createEvent
      summary: Create a new event
      description: Creates a calendar event. The authenticated user becomes the event creator.
      tags:
        - Events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCreateRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Event"
                  meta:
                    type: "null"
                  error:
                    type: "null"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /events/{id}:
    get:
      operationId: getEvent
      summary: Get a single event
      description: Returns the full details of a single event by ID.
      tags:
        - Events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Event ID
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Event"
                  meta:
                    type: "null"
                  error:
                    type: "null"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateEvent
      summary: Update an event
      description: Partially updates an existing event. Only provided fields are changed.
      tags:
        - Events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Event ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdateRequest"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Event"
                  meta:
                    type: "null"
                  error:
                    type: "null"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteEvent
      summary: Delete an event
      description: Deletes a calendar event. Only the owner or admin can delete.
      tags:
        - Events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Event ID
        - name: mode
          in: query
          schema:
            type: string
            enum: [single, future, all]
            default: all
          description: Delete mode for repeating events
      responses:
        "204":
          description: Event deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  schemas:
    ErrorResponse:
      type: object
      description: Standard error object returned in the envelope when a request fails.
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 404
        message:
          type: string
          description: Human-readable error message
          example: Event not found
        details:
          type: array
          items:
            type: string
          description: Optional list of detailed error messages (e.g. validation errors)
          default: []
    PaginationMeta:
      type: object
      description: Pagination metadata included in list responses.
      required:
        - total
        - page
        - limit
      properties:
        total:
          type: integer
          description: Total number of items matching the query
          example: 100
        page:
          type: integer
          description: Current page number (1-based)
          example: 1
        limit:
          type: integer
          description: Maximum items per page
          example: 20
...
    Event:
      type: object
      description: A calendar event.
      required:
        - id
        - title
        - start_date
        - duration
        - access
        - type
        - created_by
        - all_day
      properties:
        id:
          type: integer
          description: Unique event identifier
          example: 42
        uid:
          type: string
          description: RFC 5545 UID
        title:
          type: string
          description: Event title
          example: Team Meeting
        description:
          type: string
          description: Event description
          example: Weekly sync
        start_date:
          type: string
          pattern: '^\d{8}$'
          description: Start date (YYYYMMDD)
          example: "20260315"
        start_time:
          type: string
          pattern: '^\d{6}$'
          nullable: true
          description: Start time (HHMMSS), null for all-day
          example: "100000"
        end_date:
          type: string
          pattern: '^\d{8}$'
          description: End date (YYYYMMDD)
          example: "20260315"
        end_time:
          type: string
          pattern: '^\d{6}$'
          nullable: true
          description: End time (HHMMSS), null for all-day
          example: "110000"
        duration:
          type: integer
          description: Duration in minutes
          example: 60
        location:
          type: string
          description: Event location
          example: Room A
        access:
          type: string
          enum: [P, C, R]
          description: "Access level: P=Public, C=Confidential, R=Private"
          example: P
        type:
          type: string
          enum: [E, M, T, J, N, O]
          description: "Event type: E=Event, M=Repeating, T=Task, J=Journal, N=RepTask, O=RepJournal"
          example: E
        created_by:
          type: string
          description: Login of the user who created the event
          example: admin
        all_day:
          type: boolean
          description: Whether this is an all-day event
          example: false
        sequence:
          type: integer
          description: Change sequence number
        status:
          type: string
          nullable: true
          description: Event status
    EventCreateRequest:
      type: object
      description: Request body for creating a new event.
      required:
        - title
        - start_date
      properties:
        title:
          type: string
          description: Event title
          example: Team Meeting
        description:
          type: string
          description: Event description
        start_date:
          type: string
          pattern: '^\d{8}$'
          description: Start date (YYYYMMDD)
          example: "20260315"
        start_time:
          type: string
          pattern: '^\d{6}$'
          description: Start time (HHMMSS), omit for all-day
          example: "100000"
        duration:
          type: integer
          description: Duration in minutes
          default: 0
          example: 60
        location:
          type: string
          description: Event location
        access:
          type: string
          enum: [P, C, R]
          default: P
          description: Access level
        type:
          type: string
          enum: [E, M, T, J, N, O]
          default: E
          description: Event type
    EventUpdateRequest:
      type: object
      description: Request body for updating an event. All fields optional.
      properties:
        title:
          type: string
        description:
          type: string
        start_date:
          type: string
          pattern: '^\d{8}$'
        start_time:
          type: string
          pattern: '^\d{6}$'
        duration:
          type: integer
        location:
          type: string
        access:
          type: string
          enum: [P, C, R]
        type:
          type: string
          enum: [E, M, T, J, N, O]
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: "null"
              meta:
                type: "null"
              error:
                $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            meta: null
            error:
              code: 401
              message: Invalid or expired token
              details: []
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: "null"
              meta:
                type: "null"
              error:
                $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            meta: null
            error:
              code: 404
              message: Resource not found
              details: []
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: "null"
              meta:
                type: "null"
              error:
                $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            meta: null
            error:
              code: 403
              message: Forbidden
              details: []
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: "null"
              meta:
                type: "null"
              error:
                $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            meta: null
            error:
              code: 400
              message: Validation failed
              details:
                - "title is required"